ai_engine/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
numpy
pydantic
qrcode[pil]
pillow
python-multipart
python-dotenv
